voice_mental_health_assistant/core/requirements.txt
streamlit
openai-whisper
transformers
torch
torchaudio
pydub
